Add handling of annotations within TraceModel
This also adds two initial annotation types: RectAnnotation and
XMarkerAnnotation. RectAnnotation draws a rectangle to highlight
a certain area of the timeline. XMarkerAnnotation draws a vertical
line across the timeline to highlight a specific timestamp.

Functionality to load the annotations from trace file will follow
this change, then finally the ability to create and save annotations
to file.